Add bring in flag alert
We've been having some nasty weather this spring, and I don't always notice until its been going for a while. This will alert me when its raining consistently or the wind is too high.

I picked 0.05in/h for rain, but I'm sure that will be too low and I'll be editing this again in the future