Fix away mode restore issues
I was having problems where the awayMode restore scene wasn't doing its job. I'd leave a light on, leave home, come back, and the light would never be on.

The dA addon logs showed a few issues restoring the scene, which I first thought might be the culprit. But it wasn't the restore that was faulty, but rather some logic in the create.

The `triggerAwayMode()` function is called from a few onUpdate handlers; one for detecting how many entities are "home" and one for detecting if the nearest distance is outside a radius. This is fairly standard for away automations, and is explained better in [1].

Whats important here is that `triggerAwayMode()` was being called multiple times; once on the initial "away", and then again after the house was already in "away" mode. When creating scenes programatically, calling `create` on a scene that already exists will _update the existing scene_. In this case, it was overwriting the saved "home" state with the current, away state, which is "everything is off".

There are a couple fixes that you could apply here. You could check that the scene doesn't exist before the create call, and skip creating if it does exist. Or you can do what I did, and just check in the start of both `triggerAwayMode()` and `triggerHomeMode()` that we're not presently away/home, and exit early if we are.