Improve warm-chickens
Debounce was the _wrong_ thing to use. Deounce ensured that it would
actually never get called past the first time, because the temp
would change too frequently, causing it to retrigger and  reset
the timer within a five minute span. Fix was converting it to
throttle which ensures that it doesn't fire more than once every five
minute