🐔 add cron-check for door
Had some issues where the door would sometimes not open, and the
diagnostics would claim brownout. This is likely due to the cheap
h-bridge I've got sometimes failing to provide a stable power supply.
I could spend some time reengineering everything, putting in a better
power supply, all that stuff, but for now I try with a hack.

This adds a cron job, firing every hour between 9-12, that checks if
auto-open is enabled AND if the door is closed, and if so, opens it.